# bank-account-service/src/main/resources/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Bank Account Service
servers:
  - url: http://localhost:8081
paths:
  /accounts:
    post:
      tags:
        - account
      summary: Create a new bank account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankAccount"
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
    get:
      tags:
        - account
      summary: Get all account
      operationId: getAllAccounts
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
  /accounts/{accountId}:
    get:
      tags:
        - account
      summary: Get account details
      operationId: getAccountById
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
    delete:
      tags:
        - account
      summary: Delete account
      description: Delete account
      operationId: deleteAccount
      parameters:
        - name: accountId
          in: path
          description: Bank account id to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account deleted
        '400':
          description: Invalid account id
        '404':
          description: Account not found
  /transactions:
    post:
      tags:
        - transaction
      summary: Create a new transaction
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Transaction"
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
  /transactions/findAll/{accountId}:
    get:
      tags:
        - account
      summary: Get transactions for a bank account
      operationId: getAccountTransactions
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
  /credits:
    post:
      tags:
        - credit
      summary: Create a new credit
      operationId: createCredit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credit"
      responses:
        '201':
          description: Credit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
    get:
      tags:
        - credit
      summary: Get all credits
      operationId: getAllCredits
      responses:
        '200':
          description: List of credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
  /credits/{creditId}:
    get:
      tags:
        - credit
      summary: Get credit details
      operationId: getCreditById
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
    delete:
      tags:
        - credit
      summary: Delete credit
      operationId: deleteCredit
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Credit deleted
        '404':
          description: Credit not found
  /credits/client/{clientId}:
    get:
      tags:
        - credit
      summary: Get credits by client ID
      operationId: getCreditsByClientId
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of credits for a client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
        '404':
          description: No credits found for the client

  /reports/products:
    get:
      tags:
        - reports
      summary: Generate a general report of bank products
      operationId: generateProductReport
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Start date of the report (YYYY-MM-DD)
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: End date of the report (YYYY-MM-DD)
      responses:
        '200':
          description: General product report generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelApiResponse"
        '400':
          description: Invalid date range or missing parameters
        '404':
          description: No data found for the specified date range

components:
  schemas:
    BankAccount:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - savings
            - current
            - fixed_term
        balance:
          type: number
          format: double
          default: 0.00
          minimum: 0.00
        clientId:
          type: string
        commissionFree:
          type: boolean
          default: false
        commissionAmount:
          type: number
          format: double
          default: 0.00
          minimum: 0.00
        movementsFree:
          type: boolean
          default: false
        movementsLimit:
          type: integer
          default: 0

    Transaction:
      type: object
      properties:
        id:
          type: string
        accountId:
          type: string
        type:
          type: string
          enum:
            - deposit
            - withdrawal
        amount:
          type: number
          format: double
          minimum: 0
          example: 200.00
        date:
          type: string
          format: date
    Credit:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - personal
            - business
            - creditcardpersonal
            - creditcardbusiness
        clientId:
          type: string
        balance:
          type: number
          format: double
        limit:
          type: number
          format: double
        interestRate:
          type: number
          format: double
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - active
            - closed
            - overdue
        monthlyPayment:
          type: number
          format: double
        guarantor:
          type: string
    Client:
      type: object
      properties:
        id:
          type: string
          example: "843un884931019rj10932"
        name:
          type: string
          example: Rafael
        lastname:
          type: string
          example: Ponte
        address:
          type: string
          example: Lima - Perú
        phone:
          type: string
          example: '987654321'
        birthDate:
          type: string
        type:
          type: string
          enum:
            - personal
            - business
        profile:
          type: string
          enum:
            - vip
            - pyme
        hasCreditCard:
          type: boolean
          default: false
          example: true

    ProductReport:
      type: object
      properties:
        productType:
          type: string
          description: Type of the product
        totalBalance:
          type: number
          format: double
          description: Total balance for the product type
        totalAccounts:
          type: integer
          description: Total number of accounts or credits for the product type
        averageBalance:
          type: number
          format: double
          description: Average balance across accounts for the product type
        totalTransactions:
          type: integer
          description: Total number of transactions for the product type
        transactionsSummary:
          type: array
          items:
            type: object
            properties:
              transactionType:
                type: string
              totalAmount:
                type: number
                format: double
        reportDate:
          type: string

    ModelApiResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code of the response
        message:
          type: string
          description: Message for the response
        data:
          type: object
          description: Data for the response